4/18일자 수정

+++ View Model/contextController/LocalFileListCellController.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class LocalFileListCellController{
+class LocalFileListCellController:UIViewController{
     var dv:HomeDeviceCollectionVC?
     
     
@@ -30,7 +30,9 @@ class LocalFileListCellController{
             cell.lblDevice.isHidden = false
             cell.lblDevice.text = folderArray[indexPath.row].devNm
         } else {
-            cell.lblDevice.isHidden = true
+            cell.lblDevice.isHidden = false
+            let size = FileUtil().covertFileSize(getSize: folderArray[indexPath.row].fileSize)
+            cell.lblDevice.text = size
         }
         
 //        let swipeLeft = UISwipeGestureRecognizer(target: self, action: #selector(cellLocalFileSwipeToLeft(sender:)))
@@ -51,8 +53,9 @@ class LocalFileListCellController{
         return cell
     }
     func localContextMenuCalled(cell:LocalFileListCell, indexPath:IndexPath, sender:UIButton, folderArray:[App.FolderStruct], deviceName:String, parentView:String, deviceView:HomeDeviceCollectionVC, userId:String, fromOsCd:String, currentDevUuid:String, currentFolderId:String, viewState:HomeViewController.viewStateEnum, containerView:ContainerViewController){
+        dv = deviceView
         let fileNm = folderArray[indexPath.row].fileNm
-//        let etsionNm = folderArray[indexPath.row].etsionNm
+        let etsionNm = folderArray[indexPath.row].etsionNm
         let amdDate = folderArray[indexPath.row].amdDate
         let foldrWholePathNm = folderArray[indexPath.row].foldrWholePathNm
         let fileId = String(folderArray[indexPath.row].fileId)
@@ -61,16 +64,14 @@ class LocalFileListCellController{
         switch sender {
         case cell.btnShow:
             print("fileId: \(fileId)")
-            deviceView.showLocalFileOption(tag: sender.tag)
+            self.dv?.hideSelectedOptions(tag: sender.tag)
             
             let fileIdDict = ["fileId":fileId,"foldrWholePathNm":foldrWholePathNm,"deviceName":deviceName]
             NotificationCenter.default.post(name: Notification.Name("getFileIdFromBtnShow"), object: self, userInfo: fileIdDict)
             
             break
         case cell.btnAction:
-            if parentView == "device" {
-                deviceView.showLocalFileOption(tag: sender.tag)
-            }
+            self.dv?.hideSelectedOptions(tag: sender.tag)
             let url:URL = FileUtil().getFileUrl(fileNm: fileNm, amdDate: amdDate)!
             let urlDict = ["url":url]
             NotificationCenter.default.post(name: Notification.Name("openDocument"), object: self, userInfo: urlDict)
@@ -79,11 +80,9 @@ class LocalFileListCellController{
             break
             
         case cell.btnNas:
-//            if parentView == "device" {
-//                deviceView.showLocalFileOption(tag: sender.tag)
-//            }
-            deviceView.showLocalFileOption(tag: sender.tag)
-            let fileDict = ["fileId":fileId, "fileNm":fileNm,"amdDate":amdDate, "oldFoldrWholePathNm":foldrWholePathNm,"toStorage":"nas","fromUserId":userId, "fromOsCd":fromOsCd,"fromDevUuid":currentDevUuid]
+            self.dv?.hideSelectedOptions(tag: sender.tag)
+            let fileDict = ["fileId":fileId, "fileNm":fileNm,"amdDate":amdDate, "oldFoldrWholePathNm":foldrWholePathNm,"toStorage":"nas","fromUserId":userId, "fromOsCd":fromOsCd,"fromDevUuid":currentDevUuid,"etsionNm":etsionNm]
+            
             print("fileDict : \(fileDict)")
             
             NotificationCenter.default.post(name: Notification.Name("nasFolderSelectSegue"), object: self, userInfo: fileDict)
@@ -93,10 +92,11 @@ class LocalFileListCellController{
             
         case cell.btnGDrive:
 //            if parentView == "device" {
-                deviceView.showLocalFileOption(tag: sender.tag)
+                self.dv?.hideSelectedOptions(tag: sender.tag)
                 
                 
                 let fileDict = ["fileId":fileId, "fileNm":fileNm,"amdDate":amdDate, "oldFoldrWholePathNm":foldrWholePathNm,"fromDevUuid":currentDevUuid, "toStorage":"googleDrive","fromUserId":userId, "fromOsCd":fromOsCd]
+                
                 print("fileDict : \(fileDict)")
                 
 //                deviceView.googleSignInCheck(name: fileNm, path: foldrWholePathNm, fileDict: fileDict)
@@ -105,33 +105,41 @@ class LocalFileListCellController{
             break
         case cell.btnDelete:
 //            if parentView == "device" {
-            deviceView.showLocalFileOption(tag: sender.tag)
+            self.dv?.hideSelectedOptions(tag: sender.tag)
 //            }
             let alertController = UIAlertController(title: nil, message: "해당 파일을 삭제 하시겠습니까?", preferredStyle: .alert)
             let yesAction = UIAlertAction(title: "확인", style: UIAlertActionStyle.default) {
                 UIAlertAction in
-                let pathForRemove:String = FileUtil().getFilePath(fileNm: fileNm, amdDate: amdDate)
-                print("pathForRemove : \(pathForRemove)")
-                self.removeFile(path: pathForRemove)
-                SyncLocalFilleToNas().sync(view: "", getFoldrId: "")
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.5, execute: {
-                    let alertController = UIAlertController(title: nil, message: "파일 삭제가 완료 되였습니다.", preferredStyle: .alert)
-                    let yesAction = UIKit.UIAlertAction(title: "확인", style: UIAlertActionStyle.default) {
-                        UIAlertAction in
-                        if(viewState == .search){
-                            deviceView.refreshSearchList()
-                        } else {
-                            let fileDict = ["foldrId":String(foldrId)]
-                            print("delete filedict : \(fileDict)")
-                            NotificationCenter.default.post(name: Notification.Name("refreshInsideList"), object: self, userInfo: fileDict)                            
+                NotificationCenter.default.post(name: Notification.Name("homeViewToggleIndicator"), object: self, userInfo: nil)
+                
+                print("fileNm : \(fileNm)")
+                if let pathForRemove:String = FileUtil().getFilePathWithFoldr(fileNm: fileNm, foldrWholePathNm:foldrWholePathNm, amdDate: amdDate)
+                {
+                    print("pathForRemove : \(pathForRemove)")
+                    self.removeFile(path: pathForRemove)
+                    SyncLocalFilleToNas().sync(view: "", getFoldrId: "")
+                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.5, execute: {
+                        let alertController = UIAlertController(title: nil, message: "파일 삭제가 완료 되였습니다.", preferredStyle: .alert)
+                        let yesAction = UIKit.UIAlertAction(title: "확인", style: UIAlertActionStyle.default) {
+                            UIAlertAction in
+                            if(viewState == .search){
+                                deviceView.refreshSearchList()
+                                NotificationCenter.default.post(name: Notification.Name("homeViewToggleIndicator"), object: self, userInfo: nil)
+                            } else {
+                                let fileDict = ["foldrId":String(foldrId)]
+                                print("delete filedict : \(fileDict)")
+                                NotificationCenter.default.post(name: Notification.Name("refreshInsideList"), object: self, userInfo: fileDict)
+                                NotificationCenter.default.post(name: Notification.Name("homeViewToggleIndicator"), object: self, userInfo: nil)
+                                
+                            }
+                            
                             
                         }
-                        
-                        
-                    }
-                    alertController.addAction(yesAction)
-                    deviceView.present(alertController, animated: true)
-                })
+                        alertController.addAction(yesAction)
+                        deviceView.present(alertController, animated: true)
+                    })
+                }
+               
                 
             }
             let noAction = UIAlertAction(title: "취소", style: UIAlertActionStyle.cancel)
@@ -147,11 +155,12 @@ class LocalFileListCellController{
     
     func localContextMenuCalledFromGrid(indexPath:IndexPath, fileId:String, foldrWholePathNm:String, deviceName:String, parentView:String, deviceView:HomeViewController, userId:String, fromOsCd:String, currentDevUuid:String, currentFolderId:String, folderArray:[App.FolderStruct], intFolderArrayIndexPathRow:Int, containerView:ContainerViewController){
         let fileNm = folderArray[intFolderArrayIndexPathRow].fileNm
-//        let etsionNm = folderArray[intFolderArrayIndexPathRow].etsionNm
+        let etsionNm = folderArray[intFolderArrayIndexPathRow].etsionNm
         let amdDate = folderArray[intFolderArrayIndexPathRow].amdDate
         let foldrWholePathNm = folderArray[intFolderArrayIndexPathRow].foldrWholePathNm
         let fileId = String(folderArray[intFolderArrayIndexPathRow].fileId)
         let foldrId = folderArray[intFolderArrayIndexPathRow].foldrId
+        
         
         switch indexPath.row {
         case 0:
@@ -160,28 +169,28 @@ class LocalFileListCellController{
             NotificationCenter.default.post(name: Notification.Name("getFileIdFromBtnShow"), object: self, userInfo: fileIdDict)
             break
         case 1:
+            NotificationCenter.default.post(name: Notification.Name("toggleBottomMenu"), object: self)
             let url:URL = FileUtil().getFileUrl(fileNm: fileNm, amdDate: amdDate)!
             let urlDict = ["url":url]
             NotificationCenter.default.post(name: Notification.Name("openDocument"), object: self, userInfo: urlDict)
             
+            print("btnActino called")
 //            print("btnActino called")
             break
             
         case 2:
             NotificationCenter.default.post(name: Notification.Name("toggleBottomMenu"), object: self)
-      
-            let fileDict = ["fileId":fileId, "fileNm":fileNm,"amdDate":amdDate, "oldFoldrWholePathNm":foldrWholePathNm,"fromDevUuid":currentDevUuid, "toStorage":"googleDrive","fromUserId":userId, "fromOsCd":fromOsCd]
-            
-//            NotificationCenter.default.post(name: Notification.Name("nasFolderSelectSegue"), object: self, userInfo: fileDict)
-            containerView.googleSignInSegueState = .loginForSend
-            containerView.googleSignInCheck(name: fileNm, path: foldrWholePathNm, fileDict: fileDict)
+            let fileDict = ["fileId":fileId, "fileNm":fileNm,"amdDate":amdDate, "oldFoldrWholePathNm":foldrWholePathNm,"toStorage":"nas","fromUserId":userId, "fromOsCd":fromOsCd,"fromDevUuid":currentDevUuid,"etsionNm":etsionNm]
+            print("fileDict : \(fileDict)")
+            NotificationCenter.default.post(name: Notification.Name("nasFolderSelectSegue"), object: self, userInfo: fileDict)
             
             break
         case 3:
             let fileDict = ["fileId":fileId, "fileNm":fileNm,"amdDate":amdDate, "oldFoldrWholePathNm":foldrWholePathNm,"fromDevUuid":currentDevUuid, "toStorage":"googleDrive","fromUserId":userId, "fromOsCd":fromOsCd]
 //            print("fileDict : \(fileDict)")
-            GoogleWork().googleSignInCheck(name: fileNm, path: foldrWholePathNm, fileDict: fileDict)
+//            GoogleWork().googleSignInCheck(name: fileNm, path: foldrWholePathNm, fileDict: fileDict)
             NotificationCenter.default.post(name: Notification.Name("toggleBottomMenu"), object: self)
+            containerView.googleSignInCheck(name: fileNm, path: foldrWholePathNm, fileDict: fileDict)
             break
             
         case 4:
@@ -189,9 +198,9 @@ class LocalFileListCellController{
             let yesAction = UIAlertAction(title: "확인", style: UIAlertActionStyle.default) {
                 UIAlertAction in
                 NotificationCenter.default.post(name: Notification.Name("toggleBottomMenu"), object: self)
-              
-                let pathForRemove:String = FileUtil().getFilePath(fileNm: fileNm, amdDate: amdDate)
-//                print("pathForRemove : \(pathForRemove)")
+                print("foldrWholePathNm : \(foldrWholePathNm)")
+                let pathForRemove:String = FileUtil().getFilePathWithFoldr(fileNm: fileNm, foldrWholePathNm:foldrWholePathNm, amdDate: amdDate)
+                print("pathForRemove : \(pathForRemove)")
                 self.removeFile(path: pathForRemove)
                 SyncLocalFilleToNas().sync(view: "", getFoldrId: "")
                 DispatchQueue.main.asyncAfter(deadline: .now() + 0.5, execute: {
@@ -223,6 +232,8 @@ class LocalFileListCellController{
         }
         
     }
+    
+    
     func removeFile(path:String){
         let fileManager = FileManager.default
         do {
